+++ rules/tamper/phpbb.py
@@ -1,0 +1,20 @@
+# -*- coding: utf-8 -*-
+
+"""
+    phpbb3
+    ~~~~
+
+    tamper for phpbb3 
+
+    :author:    LoRexxar <LoRexxar@gmail.com>
+    :homepage:  https://github.com/LoRexxar/Kunlun-M
+    :license:   MIT, see LICENSE for more details.
+    :copyright: Copyright (c) 2017 LoRexxar. All rights reserved
+"""
+
+phpbb = {
+}
+
+phpbb_controlled = [
+    "$request",
+]