+++ rules/tamper/roundcube.py
@@ -1,0 +1,21 @@
+#!/usr/bin/env python
+# encoding: utf-8
+'''
+@author: LoRexxar
+@contact: lorexxar@gmail.com
+@file: roundcube.py
+@time: 2020/5/7 18:27
+@desc:
+'''
+
+roundcube = {
+    "show": [1000, 10001, 10002],
+    "Q": [1000, 10001, 1009, 10002],
+    "login": [1000, 10001, 10002],
+}
+
+roundcube_controlled = [
+    "get_input_value",
+    "rcube_utils::get_input_value",
+    "getprop",
+]