+++ utils/utils.py
@@ -683,8 +683,6 @@ def get_mainstr_from_filename(filename):
     mainstr = mainstr[-1] if mainstr[-1] else mainstr[-2]
     mainstr = mainstr.split('.')[0].strip("")
 
-    s = ScanTask.objects.filter(task_name=mainstr)
-
     return mainstr